SuccessStories: Refactored Fragments into Fragment directory
