+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/Fragments/JobSiteFragment.java
@@ -14,7 +14,6 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.example.jobs4smcyouth.JobSiteFragmentRVAdapter;
-import com.example.jobs4smcyouth.JobSiteWebViewFragment;
 import com.example.jobs4smcyouth.R;
 
 import java.util.ArrayList;
