+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/Fragments/LaborLawsWebviewFragment.java
@@ -1,4 +1,4 @@
-package com.example.jobs4smcyouth;
+package com.example.jobs4smcyouth.Fragments;
 
 
 import android.graphics.Bitmap;
