+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/Fragments/RequirementFragment.java
@@ -10,7 +10,6 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
-import com.example.jobs4smcyouth.LaborLawsWebviewFragment;
 import com.example.jobs4smcyouth.R;
