+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_labor_laws_webview.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LaborLawsWebviewFragment">
+    tools:context=".Fragments.LaborLawsWebviewFragment">
 
     <WebView
         android:id="@+id/laborLaws_webView_id"
